+++ src/nds/arm9.h
@@ -168,7 +168,7 @@ class NTR_ARM9
 	u32 debug_code;
 	u32 debug_cycles;
 
-	u32 sync_cycles;
+	s16 sync_cycles;
 	bool re_sync;
 
 	NTR_MMU* mem;
